# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_daidat02_server_internal_domain_inventory.CreateDocumentRequest:
    properties:
      created_by:
        type: string
      document_code:
        type: string
      items:
        items:
          $ref: '#/definitions/github_com_daidat02_server_internal_domain_inventory.DocumentItemInput'
        type: array
      note:
        type: string
      order_id:
        type: string
      reason:
        type: string
      supplier_id:
        type: string
      type:
        type: string
    type: object
  github_com_daidat02_server_internal_domain_inventory.DocumentItemInput:
    properties:
      product_id:
        type: string
      quantity:
        type: number
      unit:
        type: string
      unit_price:
        type: number
    type: object
...
  github_com_daidat02_server_pkg_response.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
  github_com_daidat02_server_pkg_response.SuccessResponse:
    properties:
      access_token:
        type: string
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: 'API quản lý salon: user, nhân viên, khách hàng.'
  title: Salon Management API
  version: "1.0"
paths:
...
  /inventory/:
    get:
      parameters:
      - description: Tìm theo mã phiếu, tên NCC, mã đơn, tên/SĐT người tạo
        in: query
        name: search
        type: string
      - description: Sắp xếp
        enum:
        - total_amount_asc
        - total_amount_desc
        - created_at_asc
        - created_at_desc
        in: query
        name: sort
        type: string
      - description: Lọc theo loại phiếu
        enum:
        - import
        - export
        - adjust
        - sale
        - return
        in: query
        name: filter
        type: string
      - description: Trang (mặc định 1)
        in: query
        name: page
        type: integer
      - description: Số dòng/trang (mặc định 20, tối đa 100)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Danh sách phiếu kho kèm phân trang
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.SuccessResponse'
        "400":
          description: Thiếu organization
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
        "401":
          description: Thiếu token / token không hợp lệ
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy danh sách phiếu kho (tìm kiếm + lọc loại + sắp xếp + phân trang)
      tags:
      - Inventory
  /inventory/{document_id}:
    get:
      parameters:
      - description: ID phiếu kho
        in: path
        name: document_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Chi tiết phiếu kho
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.SuccessResponse'
        "400":
          description: Thiếu organization
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
        "401":
          description: Thiếu token / token không hợp lệ
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
        "404":
          description: Phiếu kho không tồn tại
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy chi tiết phiếu kho theo ID (kèm danh sách dòng sản phẩm)
      tags:
      - Inventory
  /inventory/documents:
    post:
      consumes:
      - application/json
      parameters:
      - description: Thông tin phiếu kho kèm danh sách sản phẩm
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_daidat02_server_internal_domain_inventory.CreateDocumentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tạo phiếu kho thành công
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.SuccessResponse'
        "400":
          description: Dữ liệu không hợp lệ / tồn không đủ
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
        "401":
          description: Thiếu token / token không hợp lệ
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tạo phiếu kho (header + chi tiết + ghi sổ tồn)
      tags:
      - Inventory
  /inventory/low-stock:
    get:
      parameters:
      - description: Trang (mặc định 1)
        in: query
        name: page
        type: integer
      - description: Số dòng/trang (mặc định 20, tối đa 100)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Danh sách tồn thấp kèm phân trang
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.SuccessResponse'
        "400":
          description: Thiếu organization
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
        "401":
          description: Thiếu token / token không hợp lệ
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy danh sách sản phẩm tồn thấp
      tags:
      - Inventory
  /inventory/transactions:
    get:
      parameters:
      - description: Lọc theo ID sản phẩm
        in: query
        name: product_id
        type: string
      - description: Trang (mặc định 1)
        in: query
        name: page
        type: integer
      - description: Số dòng/trang (mặc định 20, tối đa 100)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lịch sử kho kèm phân trang
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.SuccessResponse'
        "400":
          description: Thiếu organization
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
        "401":
          description: Thiếu token / token không hợp lệ
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy lịch sử kho (phân trang, lọc theo sản phẩm)
      tags:
      - Inventory
  /inventory/transactions/{product_id}:
    get:
      parameters:
      - description: ID sản phẩm
        in: path
        name: product_id
        required: true
        type: string
      - description: Sắp xếp
        enum:
        - created_desc
        - created_asc
        in: query
        name: sort
        type: string
      - description: Lọc theo loại giao dịch
        enum:
        - import
        - export
        - adjust
        - sale
        - return
        in: query
        name: inventory_type
        type: string
      - description: Lọc theo ngày (YYYY-MM-DD)
        in: query
        name: date
        type: string
      - description: Trang (mặc định 1)
        in: query
        name: page
        type: integer
      - description: Số dòng/trang (mặc định 20, tối đa 100)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lịch sử kho theo sản phẩm kèm phân trang
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.SuccessResponse'
        "400":
          description: Thiếu organization
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
        "401":
          description: Thiếu token / token không hợp lệ
          schema:
            $ref: '#/definitions/github_com_daidat02_server_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy lịch sử kho theo sản phẩm (lọc loại + ngày + sắp xếp + phân trang)
      tags:
      - Inventory
...
securityDefinitions:
  BearerAuth:
    description: '"Bearer {access_token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"